Uses only Groovy source code path
Using Java and Groovy source code aside, causes some nasty
side effects, especially as Maven uses static compilation.
We should just use the Groovy source code path and let
the Groovy compiler do the job. Native Java code can be written
anyways.
