--- cookiecutters/portlet/{{ cookiecutter.artifact_id }}/src/main/groovy/life.qbic.portal.portlet/SampleClass.groovy
@@ -1,12 +1,0 @@
-package life.qbic.portal.portlet
-
-import groovy.util.logging.Log4j2
-
-@Log4j2
-class SampleClass {
-
-    SampleClass() {
-        log.info("SampleClass object created.")
-    }
-
-}
